refactor: extract fixture external data fetching to getExternalInfo
